Create new MGSetup page from design. Setup UI experiment to switch between new component and old.
